Removing undo command and --old flag
This tool isn't going to work well with environment variables, since they're set by the environment.
It would also be insecure to run system commands from this tool itself.

To avoid this, pad-path won't change the environment variable itself, it prints the new path so that the user can save it as needed.

Also removing --old flag from CLI