Fixing testing issues when cleaning paths
When building this as a Nix package, the file system doesn't have a `/usr/` path.
So when `make_abs_path()` is called, it first tries to run `std::fs::canonicalize()` to standardize the path.
But `canonicalize()` only properly resolves relative links if that path exists in the file system.
For any file system that doesn't contain `/usr/`, this test fails, which is what happened here.
This prevented the tests from passing when being built as a Nix package.

To fix this, I switched to one of the paths that is present under Nix that are also present under other Linux OS's, namely `/bin/` and `/etc/`.
This means that `nix build` now runs correctly and a derivation is succesfully built.